Move choice classes to different file
